Move Config types into Config.ts.
